% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowFUN.R
\name{arrowFUN}
\alias{arrowFUN}
\title{internal functions used in genesGrob to produce arrows}
\usage{
arrowFUN(where, narr = 80, gr)
}
\arguments{
\item{where}{GRanges object specifying chromosome, start and end position of the target region.}

\item{narr}{Number of arrows on xrange}

\item{gr}{GRanges object specifying the location of one transcript (start to
end) and the y position as produced by ycoordFUN}
}
\value{
matrix A matrix comprising start, end, and y positions of arrows
}
\description{
internal functions used in genesGrob to produce arrows
}
\examples{
where <- GRanges(Rle('chr1'), IRanges(start = 209789729, end = 210189729))
}
