% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLD.R
\name{calculateLD}
\alias{calculateLD}
\title{From SnpMatrix objects generated with import1000GData function generade LD data}
\usage{
calculateLD(lead_snps, xSnpMatrix, ...)
}
\arguments{
\item{lead_snps}{the lead_snps to be analyzed}

\item{vcf_file}{local file or url to vcf file.}

\item{panel_file}{local file or url to panel file.}
}
\value{
Object of class SnpMatrix.
}
\description{
From SnpMatrix objects generated with import1000GData function generade LD data
}
\examples{
COMPLETE EXAMPLES
}
