% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesGrob.R
\name{genesGrob}
\alias{genesGrob}
\title{grid Grob representing gene models}
\usage{
genesGrob(where, what, ...)
}
\arguments{
\item{where}{GRanges object specifying chromosome, start and end position of the target region.}

\item{what}{TxDb object, e.g.TxDb.Hsapiens.UCSC.hg19.knownGene}
}
\value{
gList
}
\description{
grid Grob representing gene models
}
\examples{
where <- GRanges(Rle('chr1'), IRanges(start = 209789729, end = 210189729))
mygenes <- genesGrob(where = where, what = TxDb.Hsapiens.UCSC.hg19.knownGene)
pushViewport(viewport)
grid.draw(mygenes)
}
