% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import1000GData.R
\name{import1000GData}
\alias{import1000GData}
\title{Import 1000 Genomes vcf chunks and convert them to SnpMatrix objects}
\usage{
import1000GData(where, which, vcf_file, panel_file, ...)
}
\arguments{
\item{where}{GRanges object specifying chromosome, start and end position of the target region.}

\item{which}{Target population(s). Something like "CEU" or c("CEU", "CHB").}

\item{vcf_file}{local file or url to vcf file.}

\item{panel_file}{local file or url to panel file.}
}
\value{
Object of class SnpMatrix.
}
\description{
Import 1000 Genomes vcf chunks and convert them to SnpMatrix objects
}
\examples{
COMPLETE EXAMPLES
}
