% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMapping.R
\name{prepareMapping}
\alias{prepareMapping}
\title{Preparing mapping data.frame in upperTriGrob}
\usage{
prepareMapping(r2, lead, r2_cutoff, mapping, color_highlights = list(common =
  "black", lead = "blue", tag = "red"))
}
\arguments{
\item{r2}{Symmetric numeric matrix with range [0,1] representing pairwise R^2 values.}

\item{lead}{Vector of SNP identifiers for which tag SNPs are identifierd and
which are be highlighted in the mapping panel.}

\item{r2_cutoff}{The cutoff value used to identify tag SNPs.}

\item{mapping}{DataFrame with columns 'pos' and 'ids' specifying SNP positions
and SNP names.}

\item{color_highlights}{List specifying highlight colors in mapping panel. Must have
common, lead and tag slots.}
}
\value{
data.frame The same as mapping but with tag and alpha columns
}
\description{
Preparing mapping data.frame in upperTriGrob
}
\examples{
color_highlights <- list(common = 'green', lead = 'yellow', tag = 'cyan')
}
