% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRasterFUN.R
\name{prepareRaster}
\alias{prepareRaster}
\title{Maps numeric values of matrix to colors}
\usage{
prepareRaster(mat, depths, colorscheme)
}
\arguments{
\item{mat}{E.g. R^2 Symmetric numeric matrix with range [0,1] representing pairwise R^2 values.}

\item{depth}{The  depth of pairwise LD to be shown.}

\item{mapping}{DataFrame with columns 'pos' and 'ids' specifying SNP positions
and SNP names.}

\item{color_highlights}{List specifying highlight colors in mapping panel. Must have
common, lead and tag slots.}
}
\value{
data.frame The same as mapping but with tag and alpha columns
}
\description{
Maps numeric values of matrix to colors
}
\examples{
color_highlights <- list(common = 'green', lead = 'yellow', tag = 'cyan')
}

