% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setYCoordsFUN}
\alias{setYCoordsFUN}
\title{internal functions used in genesGrob to specify y position of transcripts}
\arguments{
\item{txs}{GRanges or IRanges object specifying start and end postitions of transcripts}
}
\value{
integer vector indicating y postions of transcripts
}
\description{
internal functions used in genesGrob to specify y position of transcripts
}
\examples{
where <- GRanges(Rle('chr1'), IRanges(start = 209789729, end = 210189729))
COMPLETE EXAMPLE
}

