% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toRawSnpMat}
\alias{toRawSnpMat}
\title{Create matrix with type Raw from matrices describing heterogeneity and presence of the alternative allele}
\arguments{
\item{ishet}{Integer matrix describing heterogeneity}

\item{hasalt}{Integer matrix describing whether or not the alternaive allele is present.}
}
\value{
matrix A matrix of encoding the genotype as type Raw. 01 homozygous reference allele, 02 heterozygous, 03 homozygous alternative allele, 04 missing and else.
}
\description{
Create matrix with type Raw from matrices describing heterogeneity and presence of the alternative allele
}
