% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperTriGrob.R
\name{upperTriGrob}
\alias{upperTriGrob}
\title{grid Grob representing the upper triangle of pairwise SNP statistics like R^2}
\usage{
upperTriGrob(r2, r2d, mapping, r2_cutoff = 0.8, lead = NULL, y = unit(5/6,
  "npc"), xscale, label, colorscheme, ...)
}
\arguments{
\item{r2}{Symmetric numeric matrix with range [0,1] representing pairwise R^2 values.}

\item{r2d}{Depth of shown results. r2d = 100 specifies for each SNP only the
adjacent 100 pairwise statistics should be shown.}

\item{mapping}{DataFrame with columns 'pos' and 'ids' specifying SNP positions
and SNP names.}

\item{r2_cutoff}{The cutoff value used to identify tag SNPs.}

\item{lead}{Vector of SNP identifiers for which tag SNPs are identifierd and
which are be highlighted in the mapping panel.}

\item{label}{A string which will be plotted in the upper left, e.g. 'CEU'.}
}
\value{
gList
}
\description{
grid Grob representing the upper triangle of pairwise SNP statistics like R^2
}
\examples{
where <- GRanges(Rle('chr1'), IRanges(start = 209789729, end = 210189729)
mygenes <- genesGrob(where = where, what = TxDb.Hsapiens.UCSC.hg19.knownGene)
pushViewport(viewport)
grid.draw(mygenes)
}
